Update C01-tendencies-in-content-management.md
May be its me, but the original paragraph was difficult to read. It moves backwards and forward in time too much, i think. In this proposal, I jjust moved the present time (the failure to adopt and declining of D8) to the top of the paragraph and leave the reasons for that (and for the title) with the history following. It is just a possibility, and again, it could be me (or my non-native english).